Move from double to single-precision float

+++ include/upmem.h
@@ -25,7 +25,7 @@ extern int upmem_initialized;
 
 void free_dpus();
 void init_dpus();
-void multiply_matrix_upmem(const double *A, const double *B, double *C, int rows_a, int cols_a, int cols_b);
-void process_tile_upmem(const double *A, const double *B, double *C, int rows_a, int cols_a, int cols_b);
+void multiply_matrix_upmem(const float *A, const float *B, float *C, int rows_a, int cols_a, int cols_b);
+void process_tile_upmem(const float *A, const float *B, float *C, int rows_a, int cols_a, int cols_b);
 
 #endif
