+++ src/host/init_neuron.c
@@ -9,20 +9,20 @@ NEURON *init_neuron(int num_weights)
     
     n->num_weights = num_weights;
 
-    n->w    = (double *) malloc (sizeof(double) * n->num_weights);
+    n->w    = (float *) malloc (sizeof(float) * n->num_weights);
     if(!n->w) {
         free(n);
         return NULL;
     }
 
-    n->lw   = (double *) malloc (sizeof(double) * n->num_weights);
+    n->lw   = (float *) malloc (sizeof(float) * n->num_weights);
     if(!n->lw) {
         free(n->w);
         free(n);
         return NULL;
     }
 
-    n->batch_dw = (double *) malloc (sizeof(double) * n->num_weights);
+    n->batch_dw = (float *) malloc (sizeof(float) * n->num_weights);
     if(!n->batch_dw) {
         free(n->lw);
         free(n->w);
@@ -30,11 +30,11 @@ NEURON *init_neuron(int num_weights)
         return NULL;
     }
 
-    double limit = 1.0/sqrt((double) num_weights);
+    float limit = 1.0/sqrt((float) num_weights);
 
     for(int i=0; i<num_weights; i++)
     {
-        double rand_unit = (double)rand() / (double)RAND_MAX;
+        float rand_unit = (float)rand() / (float)RAND_MAX;
         n->w[i] = (rand_unit * 2.0 - 1.0) * limit;
         n->lw[i] = n->w[i];
         n->batch_dw[i] = 0;
